reorganize src layout
Prefix internal modules with an underscore.
