+++ tsconfig.json
@@ -16,18 +16,18 @@
 	],
 	"files": [
 		"./typings/tsd.d.ts",
+		"./src/_example.ts",
+		"./src/_history/alias-ambient-history.ts",
+		"./src/_history/hashHistory.ts",
+		"./src/_history/interfaces.ts",
+		"./src/_history/memoryHistory.ts",
+		"./src/_history/stateHistory.ts",
+		"./src/_path.ts",
 		"./src/createRoute.ts",
 		"./src/createRouter.ts",
-		"./src/example.ts",
 		"./src/history.ts",
-		"./src/history/_browser.ts",
-		"./src/history/hashHistory.ts",
-		"./src/history/interfaces.ts",
-		"./src/history/memoryHistory.ts",
-		"./src/history/stateHistory.ts",
 		"./src/interfaces.d.ts",
 		"./src/main.ts",
-		"./src/util/path.ts",
 		"./tests/functional/all.ts",
 		"./tests/intern-local.ts",
 		"./tests/intern-saucelabs.ts",
